Add files via upload
In this program, the Mammal.Java is a parent, superclass to the class Human.Java. 
Human.Java is a parent to the two classes Student.Java and Doctor.Java. 

The program creates two objects, a student and a doctor. These objects inherit the features of the mammal and human classes.